added basic docker stuff

+++ server/resolvers.js
@@ -6,8 +6,15 @@ import { ObjectId } from "mongodb";
 import axios from "axios";
 import { users as usersCollection } from "./config/mongoCollections.js";
 import bcrypt from 'bcrypt';
-const client = redis.createClient();
-await client.connect();
+
+const DEBUG = true;
+
+let client = null
+
+if(!DEBUG){
+    client = redis.createClient();
+    await client.connect();
+}
 
 /**
  * Check if a value exists in the cache
@@ -15,42 +22,55 @@ await client.connect();
  * @returns {object} - The value if it exists, otherwise null
  */
 const checkCache = async (key) => {
-  const cache = await client.get(`${key}`);
-
-  if (cache) {
-    const data = await client.get(`${key}`);
-    console.log(`getting ${key} from cache`);
-    return JSON.parse(data);
-  } else {
-    return null;
+    if (!DEBUG){
+      const cache = await client.get(`${key}`);
+  
+      if (cache) {
+        const data = await client.get(`${key}`);
+        console.log(`getting ${key} from cache`);
+        return JSON.parse(data);
+      } else {
+        return null;
+      }
+    }
+    else{
+      return null
+    }
+  };
+  
+  /**
+   * Add a value to the cache
+   * @param {string} key - The key to store the value under
+   * @param {object} value - The value to store
+   */
+  const addToCache = async (key, value, exp) => {
+    if(!DEBUG){
+      // await removeFromCache(key);
+      const data = JSON.stringify(value);
+  
+      if (exp === undefined) await client.set(`${key}`, data);
+      else await client.set(`${key}`, data, "EX", exp);
+    }
+  };
+  
+  const removeFromCache = async (key) => {
+    if(!DEBUG) {
+      console.log(`removing ${key} from cache`);
+      await client.del(`${key}`);
+    }
+  };
+  
+  const clearUserCache = async  (_id) => {
+    if(!DEBUG) {
+        const keys = await client.lRange(_id, 0, -1);
+        for (const key of keys) {
+        await removeFromCache(key);
+        }
+        await removeFromCache(_id);
+    }
   }
-};
 
-/**
- * Add a value to the cache
- * @param {string} key - The key to store the value under
- * @param {object} value - The value to store
- */
-const addToCache = async (key, value, exp) => {
-  // await removeFromCache(key);
-  const data = JSON.stringify(value);
 
-  if (exp === undefined) await client.set(`${key}`, data);
-  else await client.set(`${key}`, data, "EX", exp);
-};
-
-const removeFromCache = async (key) => {
-  console.log(`removing ${key} from cache`);
-  await client.del(`${key}`);
-};
-
-const clearUserCache = async  (_id) => {
-  const keys = await client.lRange(_id, 0, -1);
-  for (const key of keys) {
-    await removeFromCache(key);
-  }
-  await removeFromCache(_id);
-}
 const getRecords = async (collection, key) => {
   let records = await checkCache(key);
